requirements.txt
numpy
opencv-python-headless==4.12.0.88
streamlit
requests
psycopg2-binary
python-dotenv
pandas